Change global_counter_ to 32-bit
The counter only uses 22 bits, so a 32-bit atomic is sufficient.
This reduces memory footprint and more accurately reflects the data.

+++ include/msg_registry.hpp
@@ -68,7 +68,7 @@ namespace ao {
 class IdGen {
 private:
     std::atomic<int64_t> last_timestamp_ms_{0};
-    std::atomic<int64_t> global_counter_{0};  // Continuous counter across time periods
+    std::atomic<int32_t> global_counter_{0};  // Continuous counter across time periods
 
     // Counter uses 22 bits (0 to 4,194,303)
     static constexpr int32_t COUNTER_BITS = 22;
